# Add more return types for visitor functions