The visitor return value names are not final and yet to discussed. I will update the documentation accordingly.
I created a table for comparing the benchmarks. There is a slight performance hit (5-10%). The number of memory allocations are unaffected.
```
julia> g = loadsnap(:soc_slashdot0902_u)
{82168, 582533} undirected simple Int64 graph

julia> dg = loadsnap(:ego_twitter_d)
{81306, 1768149} directed simple Int64 graph
```

**BENCHMARKS**
| Test                                                  | New time   | Old time   |
|-------------------------------------------------------|------------|------------|
| connected_components(g, DFS())                        | 28.892 ms  | 26.824 ms  |
| connected_components(dg, Kosaraju())                  | 55.429 ms  | 50.502 ms  |
| connected_components(dg, Tarjan())                    | 36.570 ms  | 34.143 ms  |
| distances(g, ss, DepthFirst())                        | 27.230 ms  | 25.082 ms  |
| distances(g, ss, BreadthFirst())                      | 10.846 ms  | 9.662 ms   |
| distances(g, ss, ThreadedBreadthFirst())              | 3.117 ms   | 2.975 ms   |
| parents(g, ss, DepthFirst())                          | 26.929 ms  | 24.852 ms  |
| parents(g, ss, BreadthFirst())                        | 10.449 ms  | 9.654 ms   |
| parents(g, ss, ThreadedBreadthFirst())                | 3.125 ms   | 2.989 ms   |
| distances(dg, ss, DepthFirst())                       | 25.672 ms  | 23.866 ms  |
| distances(dg, ss, BreadthFirst())                     | 15.784 ms  | 14.215 ms  |
| distances(dg, ss, ThreadedBreadthFirst())             | 4.767 ms   | 4.526 ms   |
| parents(dg, ss, DepthFirst())                         | 25.535 ms  | 23.934 ms  |
| parents(dg, ss, BreadthFirst())                       | 15.784 ms  | 14.458 ms  |
| parents(dg, ss, ThreadedBreadthFirst())               | 4.783 ms   | 4.538 ms   |
| shortest_paths(g, ss, TrackingBFS())                  | 30.665 ms  | 29.654 ms  |
| shortest_paths(dg, ss, TrackingBFS())                 | 33.841 ms  | 31.971 ms  |
| ____has_path on 100 pairs of (u, v)____               |            |            |
| test_has_path(g, DepthFirst(), srcs, dsts)            | 1.315 s    | 1.233 s    |
| test_has_path(g, BreadthFirst(), srcs, dsts)          | 262.609 ms | 246.744 ms |
| test_has_path(g, ThreadedBreadthFirst(), srcs, dsts)  | 192.383 ms | 185.362 ms |
| test_has_path(dg, DepthFirst(), srcs, dsts)           | 811.993 ms | 750.751 ms |
| test_has_path(dg, BreadthFirst(), srcs, dsts)         | 349.377 ms | 318.146 ms |
| test_has_path(dg, ThreadedBreadthFirst(), srcs, dsts) | 198.299 ms | 187.514 ms |

I've also corrected some errors which I encountered on testing LightGraphs locally. Since we don't have multi-threading enabled on CI, it was unable to detect them.

Here is the benchmark code for reference:
```julia
julia> using LightGraphs

julia> using LightGraphs.Connectivity: connected_components, DFS, Kosaraju, Tarjan

julia> using LightGraphs.Traversals: distances, parents, has_path, BreadthFirst, DepthFirst, ThreadedBreadthFirst

julia> using LightGraphs.ShortestPaths: shortest_paths, TrackingBFS

julia> using SNAPDatasets, BenchmarkTools

julia> using Random

julia> Random.seed!(1);

julia> BenchmarkTools.DEFAULT_PARAMETERS.samples = 100
100

julia> Threads.nthreads()
6

julia> g = loadsnap(:soc_slashdot0902_u)
{82168, 582533} undirected simple Int64 graph

julia> dg = loadsnap(:ego_twitter_d)
{81306, 1768149} directed simple Int64 graph

julia> # wcc
       @btime connected_components(g, DFS())
  28.892 ms (30 allocations: 3.89 MiB)

julia> # scc
       @btime connected_components(dg, Kosaraju())
  55.429 ms (24917 allocations: 7.64 MiB)

julia> @btime connected_components(dg, Tarjan())
  36.570 ms (24586 allocations: 8.25 MiB)

julia> # undirected distances and parents
       ss = [1, 5, 6];

julia> @btime distances(g, ss, DepthFirst())
  27.230 ms (20 allocations: 1.64 MiB)

julia> @btime distances(g, ss, BreadthFirst())
  10.846 ms (9 allocations: 1.89 MiB)

julia> @btime distances(g, ss, ThreadedBreadthFirst())
  3.117 ms (443 allocations: 2.00 MiB)

julia> @btime parents(g, ss, DepthFirst())
  26.929 ms (20 allocations: 1.64 MiB)

julia> @btime parents(g, ss, BreadthFirst())
  10.449 ms (9 allocations: 1.89 MiB)

julia> @btime parents(g, ss, ThreadedBreadthFirst())
  3.125 ms (443 allocations: 2.00 MiB)

julia> # directed distances and parents
       ss = [1, 7, 10];

julia> @btime distances(dg, ss, DepthFirst())
  25.672 ms (20 allocations: 1.63 MiB)

julia> @btime distances(dg, ss, BreadthFirst())
  15.784 ms (9 allocations: 1.87 MiB)

julia> @btime distances(dg, ss, ThreadedBreadthFirst())
  4.767 ms (401 allocations: 1.98 MiB)

julia> @btime parents(dg, ss, DepthFirst())
  25.535 ms (20 allocations: 1.63 MiB)

julia> @btime parents(dg, ss, BreadthFirst())
  15.883 ms (9 allocations: 1.87 MiB)

julia> @btime parents(dg, ss, ThreadedBreadthFirst())
  4.783 ms (399 allocations: 1.98 MiB)

julia> # shortest paths tracking bfs
       ss = [1, 50, 200];

julia> @btime shortest_paths(g, ss, TrackingBFS())
  30.665 ms (133869 allocations: 15.43 MiB)

julia> @btime shortest_paths(dg, ss, TrackingBFS())
  33.841 ms (147797 allocations: 16.69 MiB)

julia> # has path on 100 pairs of (u, v)
       function test_has_path(g, alg, srcs, dsts) where T
               for i in 1:100
                       has_path(g, srcs[i], dsts[i], alg)
               end
       end
test_has_path (generic function with 1 method)

julia> srcs = rand(1:nv(g), 100);

julia> dsts = rand(1:nv(g), 100);

julia> @btime test_has_path(g, DepthFirst(), srcs, dsts)
  1.315 s (2257 allocations: 86.50 MiB)

julia> @btime test_has_path(g, BreadthFirst(), srcs, dsts)
  262.609 ms (1200 allocations: 126.43 MiB)

julia> @btime test_has_path(g, ThreadedBreadthFirst(), srcs, dsts)
  192.383 ms (27656 allocations: 142.10 MiB)

julia> srcs = rand(1:nv(dg), 100);

julia> dsts = rand(1:nv(dg), 100);

julia> @btime test_has_path(dg, DepthFirst(), srcs, dsts)
  811.993 ms (2017 allocations: 64.35 MiB)

julia> @btime test_has_path(dg, BreadthFirst(), srcs, dsts)
  349.377 ms (1200 allocations: 125.11 MiB)

julia> @btime test_has_path(dg, ThreadedBreadthFirst(), srcs, dsts)
  198.299 ms (27254 allocations: 134.39 MiB)
```