@abhinavmehndiratta 2020-05-29
< 5% slowdown after using `@inline` and `UInt8`

**UPDATED BENCHMARKS**
| Test                                                  | New time   | Old time   |
|-------------------------------------------------------|------------|------------|
| connected_components(g, DFS())                        | 26.905 ms  | 26.824 ms  |
| connected_components(dg, Kosaraju())                  | 52.362 ms  | 50.502 ms  |
| connected_components(dg, Tarjan())                    | 35.447 ms  | 34.143 ms  |
| distances(g, ss, DepthFirst())                        | 25.109 ms  | 25.082 ms  |
| distances(g, ss, BreadthFirst())                      | 10.200 ms  | 9.662 ms   |
| distances(g, ss, ThreadedBreadthFirst())              | 2.984 ms   | 2.975 ms   |
| parents(g, ss, DepthFirst())                          | 24.950 ms  | 24.852 ms  |
| parents(g, ss, BreadthFirst())                        | 10.116 ms  | 9.654 ms   |
| parents(g, ss, ThreadedBreadthFirst())                | 3.125 ms   | 2.989 ms   |
| distances(dg, ss, DepthFirst())                       | 24.316 ms  | 23.866 ms  |
| distances(dg, ss, BreadthFirst())                     | 15.241 ms  | 14.215 ms  |
| distances(dg, ss, ThreadedBreadthFirst())             | 4.563 ms   | 4.526 ms   |
| parents(dg, ss, DepthFirst())                         | 24.416 ms  | 23.934 ms  |
| parents(dg, ss, BreadthFirst())                       | 15.266 ms  | 14.458 ms  |
| parents(dg, ss, ThreadedBreadthFirst())               | 4.563 ms   | 4.538 ms   |
| shortest_paths(g, ss, TrackingBFS())                  | 29.848 ms  | 29.654 ms  |
| shortest_paths(dg, ss, TrackingBFS())                 | 32.641 ms  | 31.971 ms  |
| ____has_path on 100 pairs of (u, v)____               |            |            |
| test_has_path(g, DepthFirst(), srcs, dsts)            | 1.275 s    | 1.233 s    |
| test_has_path(g, BreadthFirst(), srcs, dsts)          | 260.609 ms | 246.744 ms |
| test_has_path(g, ThreadedBreadthFirst(), srcs, dsts)  | 187.383 ms | 185.362 ms |
| test_has_path(dg, DepthFirst(), srcs, dsts)           | 777.993 ms | 750.751 ms |
| test_has_path(dg, BreadthFirst(), srcs, dsts)         | 334.377 ms | 318.146 ms |
| test_has_path(dg, ThreadedBreadthFirst(), srcs, dsts) | 192.529 ms | 187.514 ms |